Extract SelfService from OnboardRestController in onboard-server
