+++ neighbour-model/src/main/java/no/vegvesen/ixn/federation/repository/NeighbourRepository.java
@@ -6,6 +6,7 @@ import no.vegvesen.ixn.federation.model.*;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
+import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -15,6 +16,7 @@ public interface NeighbourRepository extends CrudRepository<Neighbour, Integer> 
 
 	Neighbour findByName(String name);
 
+	@NonNull
 	List<Neighbour> findAll();
 
 	List<Neighbour> findByCapabilities_Status(Capabilities.CapabilitiesStatus capabilitiesStatus);
