+++ neighbour-model/src/main/java/no/vegvesen/ixn/federation/repository/ServiceProviderRepository.java
@@ -2,8 +2,8 @@ package no.vegvesen.ixn.federation.repository;
 
 import no.vegvesen.ixn.federation.model.LocalSubscriptionStatus;
 import no.vegvesen.ixn.federation.model.ServiceProvider;
-import no.vegvesen.ixn.federation.model.SubscriptionRequestStatus;
 import org.springframework.data.repository.CrudRepository;
+import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -15,4 +15,6 @@ public interface ServiceProviderRepository extends CrudRepository<ServiceProvide
 
 	List<ServiceProvider> findBySubscriptions_StatusIn(LocalSubscriptionStatus ... statuses);
 
+	@NonNull
+	List<ServiceProvider> findAll();
 }
