+++ onboard-server/src/main/java/no/vegvesen/ixn/serviceprovider/OnboardRestController.java
@@ -1,13 +1,13 @@
 package no.vegvesen.ixn.serviceprovider;
 
-import no.vegvesen.ixn.federation.auth.CertService;
 import no.vegvesen.ixn.federation.api.v1_0.DataTypeApi;
+import no.vegvesen.ixn.federation.auth.CertService;
 import no.vegvesen.ixn.federation.exceptions.CapabilityPostException;
 import no.vegvesen.ixn.federation.exceptions.SubscriptionRequestException;
 import no.vegvesen.ixn.federation.model.*;
-import no.vegvesen.ixn.federation.repository.SelfRepository;
 import no.vegvesen.ixn.federation.repository.ServiceProviderRepository;
 import no.vegvesen.ixn.federation.transformer.DataTypeTransformer;
+import no.vegvesen.ixn.onboard.SelfService;
 import no.vegvesen.ixn.serviceprovider.model.LocalDataType;
 import no.vegvesen.ixn.serviceprovider.model.LocalDataTypeList;
 import no.vegvesen.ixn.serviceprovider.model.LocalSubscriptionApi;
@@ -20,15 +20,13 @@ import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.view.RedirectView;
 
-import java.time.LocalDateTime;
 import java.util.*;
-import java.util.stream.Collectors;
 
 @RestController
 public class OnboardRestController {
 
 	private final ServiceProviderRepository serviceProviderRepository;
-	private final SelfRepository selfRepository;
+	private final SelfService selfService;
 	private final CertService certService;
 	private DataTypeTransformer dataTypeTransformer = new DataTypeTransformer();
 	private Logger logger = LoggerFactory.getLogger(OnboardRestController.class);
@@ -39,10 +37,10 @@ public class OnboardRestController {
 
 	@Autowired
 	public OnboardRestController(ServiceProviderRepository serviceProviderRepository,
-								 SelfRepository selfRepository,
-								 CertService certService) {
+								 CertService certService,
+								 SelfService selfService) {
 		this.serviceProviderRepository = serviceProviderRepository;
-		this.selfRepository = selfRepository;
+		this.selfService = selfService;
 		this.certService = certService;
 	}
 
@@ -57,7 +55,7 @@ public class OnboardRestController {
 			throw new CapabilityPostException("Bad api object. The posted DataTypeApi object had no capabilities. Nothing to add.");
 		}
 
-		Self selfBeforeUpdate = fetchSelf();
+		Self selfBeforeUpdate = selfService.fetchSelf();
 		Set<DataType> currentSelfCapabilities = selfBeforeUpdate.getLocalCapabilities();
 
 		DataType newLocalCapability = dataTypeTransformer.dataTypeApiToDataType(capabilityDataType);
@@ -91,7 +89,7 @@ public class OnboardRestController {
 			}
 		}
 
-		updateSelfCapabilities(currentSelfCapabilities);
+		selfService.updateSelfCapabilities(currentSelfCapabilities, serviceProviderRepository.findAll());
 
 		logger.info("Returning updated Service Provider: {}", serviceProviderToUpdate.toString());
 		LocalDataType localDataType = null;
@@ -110,7 +108,7 @@ public class OnboardRestController {
 		logger.info("Capabilities - Received DELETE from Service Provider: {}", serviceProviderName);
 
 		// Get the representation of self - if it doesnt exist in the database call the method that creates it.
-		Self previousSelfRepresentation = fetchSelf();
+		Self previousSelfRepresentation = selfService.fetchSelf();
 		// Get previous Self capabilities to compare with the updated capabilities.
 		Set<DataType> previousSelfCapabilities = new HashSet<>(previousSelfRepresentation.getLocalCapabilities());
 
@@ -139,89 +137,13 @@ public class OnboardRestController {
 
 		// Save the updated Service Provider representation in the database.
 		serviceProviderRepository.save(serviceProviderToUpdate);
-
-		updateSelfCapabilities(previousSelfCapabilities);
+		List<ServiceProvider> serviceProviders = serviceProviderRepository.findAll();
+		selfService.updateSelfCapabilities(previousSelfCapabilities, serviceProviders);
 
 		logger.info("Updated Service Provider: {}", serviceProviderToUpdate.toString());
 		RedirectView redirectView = new RedirectView("/{serviceProviderName}/capabilities");
 		OnboardMDCUtil.removeLogVariables();
 		return redirectView;
-	}
-
-	private void updateSelfCapabilities(Set<DataType> previousCapabilities) {
-		// Recalculate Self capabilities now that we updated a Service Provider.
-		Self self = fetchSelf();
-		Iterable<ServiceProvider> serviceProviders = serviceProviderRepository.findAll();
-		Set<DataType> updatedCapabilities = calculateSelfCapabilities(serviceProviders);
-
-		// If old version and updated version of the capabilities are not equal, then we update the timestamp
-		logger.info("Previous capabilities: {}", previousCapabilities);
-		logger.info("Updated capabilities:  {}", updatedCapabilities);
-
-		if (!previousCapabilities.equals(updatedCapabilities)) {
-			logger.info("Capabilities have changed. Updating representation of self.");
-			self.setLastUpdatedLocalCapabilities(LocalDateTime.now());
-			self.setLocalCapabilities(updatedCapabilities);
-			selfRepository.save(self);
-			logger.info("Updated Self: {}", self.toString());
-		} else {
-			logger.info("Capabilities have not changed. Keeping the current representation of self.");
-		}
-	}
-
-	Set<DataType> calculateSelfCapabilities(Iterable<ServiceProvider> serviceProviders) {
-		logger.info("Calculating Self capabilities...");
-		Set<DataType> localCapabilities = new HashSet<>();
-
-		for (ServiceProvider serviceProvider : serviceProviders) {
-			logger.info("Service provider name: {}", serviceProvider.getName());
-			Set<DataType> serviceProviderCapabilities = serviceProvider.getCapabilities().getDataTypes();
-			logger.info("Service Provider capabilities: {}", serviceProviderCapabilities.toString());
-			localCapabilities.addAll(serviceProviderCapabilities);
-		}
-		logger.info("Calculated Self capabilities: {}", localCapabilities);
-		return localCapabilities;
-	}
-
-
-	private void updateSelfSubscriptions(Set<DataType> previousSubscriptions) {
-		// Get the current Self subscriptions. Recalculate the Self subscriptions now that a Service Provider has been updated.
-		Self self = fetchSelf();
-		Iterable<ServiceProvider> serviceProviders = serviceProviderRepository.findAll();
-		Set<DataType> updatedSubscriptions = calculateSelfSubscriptions(serviceProviders);
-
-		if (!previousSubscriptions.equals(updatedSubscriptions)) {
-			logger.debug("Subscriptions have changed from {} to {}. Updating representation of self.",
-					previousSubscriptions.toString(),
-					updatedSubscriptions.toString());
-			self.setLocalSubscriptions(updatedSubscriptions);
-			self.setLastUpdatedLocalSubscriptions(LocalDateTime.now());
-			selfRepository.save(self);
-			logger.info("Updated Self: {}", self.toString());
-		} else {
-			logger.debug("Subscriptions have not changed from {} to {}. No self update.",
-					previousSubscriptions.toString(),
-					updatedSubscriptions.toString());
-		}
-	}
-
-	Set<DataType> calculateSelfSubscriptions(Iterable<ServiceProvider> serviceProviders) {
-		logger.info("Calculating Self subscriptions...");
-		Set<DataType> localSubscriptions = new HashSet<>();
-
-		for (ServiceProvider serviceProvider : serviceProviders) {
-			logger.info("Service provider name: {}", serviceProvider.getName());
-			//Set<DataType> serviceProviderSubscriptions = serviceProvider.getOrCreateLocalSubscriptionRequest().getSubscriptions();
-			Set<DataType> serviceProviderSubscriptions = serviceProvider
-					.getSubscriptions()
-					.stream()
-					.map(LocalSubscription::getDataType)
-					.collect(Collectors.toSet());
-			logger.info("Service Provider Subscriptions: {}", serviceProviderSubscriptions.toString());
-			localSubscriptions.addAll(serviceProviderSubscriptions);
-		}
-		logger.info("Calculated Self subscriptions: {}", localSubscriptions.toString());
-		return localSubscriptions;
 	}
 
 	@RequestMapping(method = RequestMethod.POST, path = "/{serviceProviderName}/subscriptions")
@@ -239,7 +161,7 @@ public class OnboardRestController {
 		serviceProviderRepository.findByName(serviceProviderName);
 
 		// Get the representation of self - if it doesnt exist in the database call the method that creates it.
-		Self self = fetchSelf();
+		Self self = selfService.fetchSelf();
 		Set<DataType> previousSelfSubscriptions = new HashSet<>(self.getLocalSubscriptions());
 
 		ServiceProvider serviceProviderToUpdate = serviceProviderRepository.findByName(serviceProviderName);
@@ -256,7 +178,7 @@ public class OnboardRestController {
 		ServiceProvider saved = serviceProviderRepository.save(serviceProviderToUpdate);
 		logger.debug("Updated Service Provider: {}", saved.toString());
 
-		updateSelfSubscriptions(previousSelfSubscriptions);
+		selfService.updateSelfSubscriptions(previousSelfSubscriptions, serviceProviderRepository.findAll());
 		LocalSubscription returning = null;
 		for (LocalSubscription ls : saved.getSubscriptions()) {
 			if (ls.equals(localSubscription)) {
@@ -267,15 +189,6 @@ public class OnboardRestController {
 		return typeTransformer.transformLocalSubecriptionToLocalSubscriptionApi(returning);
 	}
 
-	// Get the self representation from the database. If it doesn't exist, create it.
-	private Self fetchSelf() {
-		Self self = selfRepository.findByName(nodeProviderName);
-		if (self == null) {
-			self = new Self(nodeProviderName);
-			return selfRepository.save(self);
-		}
-		return self;
-	}
 
 	@RequestMapping(method = RequestMethod.DELETE, path = "/{serviceProviderName}/subscriptions/{dataTypeId}")
 	public RedirectView deleteSubscription(@PathVariable String serviceProviderName, @PathVariable Integer dataTypeId) throws NotFoundException {
@@ -285,7 +198,7 @@ public class OnboardRestController {
 		logger.info("Service Provider {}, DELETE subscription {}", serviceProviderName, dataTypeId);
 
 		// Get the representation of self - if it doesnt exist in the database call the method that creates it.
-		Self self = fetchSelf();
+		Self self = selfService.fetchSelf();
 
 		Set<DataType> currentSelfSubscriptions = new HashSet<>(self.getLocalSubscriptions());
 
@@ -308,7 +221,7 @@ public class OnboardRestController {
 		ServiceProvider saved = serviceProviderRepository.save(serviceProviderToUpdate);
 		logger.debug("Updated Service Provider: {}", saved.toString());
 
-		updateSelfSubscriptions(currentSelfSubscriptions);
+		selfService.updateSelfSubscriptions(currentSelfSubscriptions, serviceProviderRepository.findAll());
 
 		RedirectView redirect = new RedirectView("/{serviceProviderName}/subscriptions/");
 		OnboardMDCUtil.removeLogVariables();
@@ -361,7 +274,7 @@ public class OnboardRestController {
 	// TODO: Remove
 	@RequestMapping(method = RequestMethod.GET, path = "/getSelfRepresentation", produces = MediaType.APPLICATION_JSON_VALUE)
 	public Self getSelfRepresentation() {
-		return fetchSelf();
+		return selfService.fetchSelf();
 	}
 
 }
