+++ onboard-server/src/test/java/no/vegvesen/ixn/serviceprovider/OnboardRestControllerTest.java
@@ -1,11 +1,11 @@
 package no.vegvesen.ixn.serviceprovider;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import no.vegvesen.ixn.federation.auth.CertService;
 import no.vegvesen.ixn.federation.api.v1_0.DataTypeApi;
 import no.vegvesen.ixn.federation.api.v1_0.Datex2DataTypeApi;
 import no.vegvesen.ixn.federation.api.v1_0.DenmDataTypeApi;
 import no.vegvesen.ixn.federation.api.v1_0.IviDataTypeApi;
+import no.vegvesen.ixn.federation.auth.CertService;
 import no.vegvesen.ixn.federation.model.*;
 import no.vegvesen.ixn.federation.repository.SelfRepository;
 import no.vegvesen.ixn.federation.repository.ServiceProviderRepository;
@@ -27,14 +27,9 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
@@ -301,63 +296,6 @@ public class OnboardRestControllerTest {
 						.content(subscriptionRequestApiToServerJson))
 				.andDo(print())
 				.andExpect(status().isForbidden());
-	}
-
-	@Test
-	void calculateSelfCapabilitiesTest() {
-
-		DataType a = getDatex("SE");
-		DataType b = getDatex("FI");
-		DataType c = getDatex("NO");
-
-		ServiceProvider firstServiceProvider = new ServiceProvider();
-		firstServiceProvider.setName("First Service Provider");
-		Capabilities firstServiceProviderCapabilities = new Capabilities(Capabilities.CapabilitiesStatus.KNOWN, Stream.of(a, b).collect(Collectors.toSet()));
-		firstServiceProvider.setCapabilities(firstServiceProviderCapabilities);
-
-		ServiceProvider secondServiceProvider = new ServiceProvider();
-		secondServiceProvider.setName("Second Service Provider");
-		Capabilities secondServiceProviderCapabilities = new Capabilities(Capabilities.CapabilitiesStatus.KNOWN, Stream.of(b, c).collect(Collectors.toSet()));
-		secondServiceProvider.setCapabilities(secondServiceProviderCapabilities);
-
-		Set<ServiceProvider> serviceProviders = Stream.of(firstServiceProvider, secondServiceProvider).collect(Collectors.toSet());
-
-		Set<DataType> selfCapabilities = onboardRestController.calculateSelfCapabilities(serviceProviders);
-
-		assertEquals(selfCapabilities.size(), 3);
-		assertTrue(selfCapabilities.containsAll(Stream.of(a, b, c).collect(Collectors.toSet())));
-	}
-
-	private DataType getDatex(String se) {
-		HashMap<String, String> datexHeaders = new HashMap<>();
-		datexHeaders.put(MessageProperty.MESSAGE_TYPE.getName(), Datex2DataTypeApi.DATEX_2);
-		datexHeaders.put(MessageProperty.ORIGINATING_COUNTRY.getName(), se);
-		return new DataType(datexHeaders);
-	}
-
-	@Test
-	void calculateSelfSubscriptionsTest() {
-
-		DataType localSubA = new DataType(1, MessageProperty.ORIGINATING_COUNTRY.getName(), "FI");
-		DataType localSubB = new DataType(1, MessageProperty.ORIGINATING_COUNTRY.getName(), "SE");
-		DataType localSubC = new DataType(1, MessageProperty.ORIGINATING_COUNTRY.getName(), "NO");
-
-		ServiceProvider firstServiceProvider = new ServiceProvider();
-		firstServiceProvider.setName("First Service Provider");
-		firstServiceProvider.addLocalSubscription(new LocalSubscription(LocalSubscriptionStatus.CREATED,localSubA));
-		firstServiceProvider.addLocalSubscription(new LocalSubscription(LocalSubscriptionStatus.CREATED,localSubB));
-
-		ServiceProvider secondServiceProvider = new ServiceProvider();
-		secondServiceProvider.setName("Second Service Provider");
-		firstServiceProvider.addLocalSubscription(new LocalSubscription(LocalSubscriptionStatus.CREATED,localSubB));
-		firstServiceProvider.addLocalSubscription(new LocalSubscription(LocalSubscriptionStatus.CREATED,localSubC));
-
-		Set<ServiceProvider> serviceProviders = Stream.of(firstServiceProvider, secondServiceProvider).collect(Collectors.toSet());
-
-		Set<DataType> selfSubscriptions = onboardRestController.calculateSelfSubscriptions(serviceProviders);
-
-		assertEquals(selfSubscriptions.size(), 3);
-		assertTrue(selfSubscriptions.containsAll(Stream.of(localSubA, localSubB, localSubC).collect(Collectors.toSet())));
 	}
 
 	@Test
