+++ onboard-service/src/test/java/no/vegvesen/ixn/onboard/SelfServiceTest.java
@@ -1,0 +1,89 @@
+package no.vegvesen.ixn.onboard;
+
+
+import no.vegvesen.ixn.federation.api.v1_0.Datex2DataTypeApi;
+import no.vegvesen.ixn.federation.model.*;
+import no.vegvesen.ixn.federation.repository.SelfRepository;
+import no.vegvesen.ixn.properties.MessageProperty;
+import org.junit.Assert;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.lang.NonNull;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+class SelfServiceTest {
+
+	@MockBean
+	SelfRepository selfRepository;
+
+	SelfService selfService = new SelfService(selfRepository, "myName");
+
+	@Test
+	void calculateSelfSubscriptionsTest() {
+
+		DataType localSubA = new DataType(1, MessageProperty.ORIGINATING_COUNTRY.getName(), "FI");
+		DataType localSubB = new DataType(1, MessageProperty.ORIGINATING_COUNTRY.getName(), "SE");
+		DataType localSubC = new DataType(1, MessageProperty.ORIGINATING_COUNTRY.getName(), "NO");
+
+		ServiceProvider firstServiceProvider = new ServiceProvider();
+		firstServiceProvider.setName("First Service Provider");
+		firstServiceProvider.addLocalSubscription(new LocalSubscription(LocalSubscriptionStatus.CREATED,localSubA));
+		firstServiceProvider.addLocalSubscription(new LocalSubscription(LocalSubscriptionStatus.CREATED,localSubB));
+
+		ServiceProvider secondServiceProvider = new ServiceProvider();
+		secondServiceProvider.setName("Second Service Provider");
+		firstServiceProvider.addLocalSubscription(new LocalSubscription(LocalSubscriptionStatus.CREATED,localSubB));
+		firstServiceProvider.addLocalSubscription(new LocalSubscription(LocalSubscriptionStatus.CREATED,localSubC));
+
+		List<ServiceProvider> serviceProviders = Stream.of(firstServiceProvider, secondServiceProvider).collect(Collectors.toList());
+
+		Set<DataType> selfSubscriptions = selfService.calculateSelfSubscriptions(serviceProviders);
+
+		assertEquals(selfSubscriptions.size(), 3);
+		assertTrue(selfSubscriptions.containsAll(Stream.of(localSubA, localSubB, localSubC).collect(Collectors.toSet())));
+	}
+
+
+	@Test
+	void calculateSelfCapabilitiesTest() {
+
+		DataType a = getDatex("SE");
+		DataType b = getDatex("FI");
+		DataType c = getDatex("NO");
+
+		ServiceProvider firstServiceProvider = new ServiceProvider();
+		firstServiceProvider.setName("First Service Provider");
+		Capabilities firstServiceProviderCapabilities = new Capabilities(Capabilities.CapabilitiesStatus.KNOWN, Stream.of(a, b).collect(Collectors.toSet()));
+		firstServiceProvider.setCapabilities(firstServiceProviderCapabilities);
+
+		ServiceProvider secondServiceProvider = new ServiceProvider();
+		secondServiceProvider.setName("Second Service Provider");
+		Capabilities secondServiceProviderCapabilities = new Capabilities(Capabilities.CapabilitiesStatus.KNOWN, Stream.of(b, c).collect(Collectors.toSet()));
+		secondServiceProvider.setCapabilities(secondServiceProviderCapabilities);
+
+		Set<ServiceProvider> serviceProviders = Stream.of(firstServiceProvider, secondServiceProvider).collect(Collectors.toSet());
+
+		Set<DataType> selfCapabilities = selfService.calculateSelfCapabilities(serviceProviders);
+
+		Assert.assertEquals(selfCapabilities.size(), 3);
+		Assert.assertTrue(selfCapabilities.containsAll(Stream.of(a, b, c).collect(Collectors.toSet())));
+	}
+
+	@NonNull
+	private DataType getDatex(String se) {
+		HashMap<String, String> datexHeaders = new HashMap<>();
+		datexHeaders.put(MessageProperty.MESSAGE_TYPE.getName(), Datex2DataTypeApi.DATEX_2);
+		datexHeaders.put(MessageProperty.ORIGINATING_COUNTRY.getName(), se);
+		return new DataType(datexHeaders);
+	}
+
+
+}
